refactor: WebSocketConfig 파일 이동 및 JwtHandshakeInterceptor 파일 생성 - #6
