+++ chat/src/main/java/com/back/chat/adapter/in/ws/JwtHandshakeInterceptor.java
@@ -1,0 +1,4 @@
+package com.back.chat.adapter.in.ws;
+
+public class JwtHandshakeInterceptor {
+}
